+++ src/VaultSharp/V1/AuthMethods/AuthMethodDefaultPaths.cs
@@ -11,6 +11,7 @@
         public const string JWT = "jwt";
         public const string Kubernetes = "kubernetes";
         public const string LDAP = "ldap";
+        public const string Kerberos = "kerberos";
         public const string Okta = "okta";
         public const string RADIUS = "radius";
         public const string Cert = "cert";