+++ src/VaultSharp/V1/SecretsEngines/PKI/RevokeCertificateResponse.cs
@@ -1,0 +1,19 @@
+﻿using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.PKI
+{
+    /// <summary>
+    /// Represents the Certificate revocation response.
+    /// </summary>
+    public class RevokeCertificateResponse
+    {
+        /// <summary>
+        /// Gets or sets the revocation time.
+        /// </summary>
+        /// <value>
+        /// The revocation time.
+        /// </value>
+        [JsonProperty("revocation_time")]
+        public int RevocationTime { get; set; }
+    }
+}