+++ src/VaultSharp/V1/SecretsEngines/SSH/ISSHSecretsEngine.cs
@@ -28,5 +28,19 @@ namespace VaultSharp.V1.SecretsEngines.SSH
         /// The secret with the SSH credentials.
         /// </returns>
         Task<Secret<SSHCredentials>> GetCredentialsAsync(string roleName, string ipAddress, string username = null, string mountPoint = SecretsEngineDefaultPaths.SSH, string wrapTimeToLive = null);
+
+        /// <summary>
+        /// This endpoint signs an SSH public key based on the supplied parameters, 
+        /// subject to the restrictions contained in the role named in the endpoint.
+        /// </summary>
+        /// <param name="roleName"><para>[required]</para>
+        /// Specifies the name of the role to sign.</param>
+        /// <param name="signKeyRequest"><para>[required]</para>
+        /// The request parameters</param>
+        /// <param name="mountPoint"><para>[optional]</para>
+        /// The mount point for the SSH backend. Defaults to <see cref="SecretsEngineDefaultPaths.SSH" />
+        /// Provide a value only if you have customized the SSH mount point.</param>
+        /// <returns>The signed key</returns>
+        Task<Secret<SignedKeyResponse>> SignKeyAsync(string roleName, SignKeyRequest signKeyRequest, string mountPoint = SecretsEngineDefaultPaths.SSH);
     }
 }